+++ src/components/InputField/index.ts
@@ -1,0 +1,2 @@
+export { InputField as default } from './InputField';
+export type { InputFieldProps } from './InputField';
