--- src/components/TextInput/TextInput.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as TextInputStoryFile from './TextInput.stories';
-
-describe('<TextInput />', () => {
-  generateSnapshots(TextInputStoryFile);
-});
