--- src/components/TextInput/index.ts
@@ -1,2 +1,0 @@
-export { TextInput as default } from './TextInput';
-export type { TextInputProps } from './TextInput';
