+++ src/index.ts
@@ -39,6 +39,7 @@ export { default as HorizontalStepper } from './components/HorizontalStepper';
 export { default as Hr } from './components/Hr';
 export { default as Icon } from './components/Icon';
 export { default as InlineNotification } from './components/InlineNotification';
+export { default as InputField } from './components/InputField';
 export { default as InputLabel } from './components/InputLabel';
 export { default as Label } from './components/Label';
 export { default as Layout } from './components/Layout';
@@ -86,7 +87,6 @@ export { default as Tabs } from './components/Tabs';
 export { default as Tag } from './components/Tag';
 export { default as Text } from './components/Text';
 export { default as TextField } from './components/TextField';
-export { default as TextInput } from './components/TextInput';
 export { default as TimelineNav } from './components/TimelineNav';
 export { default as TimelineNavPanel } from './components/TimelineNavPanel';
 export { default as Toast } from './components/Toast';
